requirements_pip.txt
# Core scientific computing - let pip resolve compatible versions
numpy>=1.17.0,<1.24.0
scipy>=1.4.0
pandas>=1.3.0
matplotlib>=3.3.0

# Image processing
scikit-image>=0.19.0
opencv-python>=4.5.0
opencv-python-headless>=4.5.0
Pillow>=8.0.0
imageio>=2.9.0
tifffile>=2021.0.0
PyWavelets>=1.1.0

# Machine learning
scikit-learn>=1.0.0,<1.3.0
joblib>=1.0.0
umap-learn>=0.5.0
pynndescent>=0.5.0
imblearn

# Deep learning - use newer tensorflow that's compatible with newer numpy
tensorflow>=2.10.0
stardist>=0.8.0
csbdeep>=0.6.0

# Statistics
statsmodels>=0.13.0
numba>=0.54.0
llvmlite>=0.37.0

# Visualization
seaborn>=0.11.0
tqdm>=4.58.0

# Utilities
packaging>=21.3
six>=1.16.0
python-dateutil>=2.8.0
pytz>=2021.0
requests>=2.26.0
urllib3>=1.26.0
certifi>=2021.0.0
networkx>=2.5

# Jupyter support
jupyter-core>=4.6.0
pyzmq>=19.0.0
tornado>=6.0.0

# CLI dependencies
pyyaml>=5.4
